--- ApiEndpointSimulator/birthdays.csv
@@ -1,1 +1,0 @@
-﻿<<lastName>>,<<firstName>>,<<dd/mm/yyyy>>
