+++ BirthdayApi/CsvParser/CsvReaderWrapper.cs
@@ -1,11 +1,11 @@
-﻿using ApiEndpointSimulator.Models;
+﻿using BirthdayApi.Models;
 using CsvHelper;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
 
-namespace ApiEndpointSimulator.CsvParser
+namespace BirthdayApi.CsvParser
 {
     public static class CsvReaderWrapper
     {
@@ -14,7 +14,7 @@ namespace ApiEndpointSimulator.CsvParser
             using (var reader = new StreamReader(filePath))
             using (var csvReader = new CsvReader(reader, new CultureInfo("EN")))
             {
-                csvReader.Configuration.HasHeaderRecord = false;
+                csvReader.Configuration.HasHeaderRecord = true;
 
                 return csvReader.GetRecords<BirthdayPerson>().ToList();
             }
