+++ BirthdayApi/Startup.cs
@@ -4,7 +4,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace ApiEndpointSimulator
+namespace BirthdayApi
 {
     public class Startup
     {
