+++ BirthdayApp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29709.97
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BirthdayApi", "ApiEndpointSimulator\BirthdayApi.csproj", "{F005E4F6-C517-47DE-8158-EFE7C6B97F85}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BirthdayApi", "BirthdayApi\BirthdayApi.csproj", "{F005E4F6-C517-47DE-8158-EFE7C6B97F85}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
